requirements.txt
certifi==2024.2.2
charset-normalizer==3.3.2
contourpy==1.2.1
cycler==0.12.1
fonttools==4.51.0
idna==3.7
kiwisolver==1.4.5
matplotlib==3.8.4
numpy==1.26.4
packaging==24.0
pandas==2.2.2
pillow==10.3.0
pycairo==1.26.0
PyGObject==3.48.2
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.31.0
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
six==1.16.0
tzdata==2024.1
urllib3==2.2.1
vext==0.7.6
vext.gi==0.7.4
wordcloud==1.9.3
